feat : 유동적인 URL 제작 성공
