+++ src/components/NavigateBar.jsx
@@ -2,47 +2,15 @@ import React, { useState, useEffect, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styled, { keyframes } from 'styled-components';
 import Swal from 'sweetalert2';
-import bell from '../assets/img/NavigateBar_bell.png';
 import arrow from '../assets/img/NavigateBar_arrow.png';
 import ProfileMenu from './CalendarPage/ProfileMenu';
 import useUserInfoStore from '../stores/userInfoStore';
 
 const NavigateBar = () => {
   const [isProfMenuOpen, setIsProfMenuOpen] = useState(false);
-  const [isNotifyMenuOpen, setIsNotifyMenuOpen] = useState(false);
   const profMenuRef = useRef(null);
-  const notifyMenuRef = useRef(null);
   const userInfoStore = useUserInfoStore();
   const { userInfoList } = userInfoStore;
-  const navigate = useNavigate();
-  const [showLoginAlert, setShowLoginAlert] = useState(false);
-
-  useEffect(() => {
-    const loggedInUserId = localStorage.getItem('loggedInUserId');
-    const loggedInUserNickname = localStorage.getItem('loggedInUserNickname');
-
-    if (loggedInUserId && loggedInUserNickname) {
-      // 기존 정보를 제거하고 새로운 정보를 추가
-      userInfoStore.removeUserInfo(loggedInUserId);
-      userInfoStore.addUserInfo(loggedInUserId, loggedInUserNickname);
-    }
-
-    // 화면이 처음 마운트될 때는 SweetAlert 창을 띄우지 않도록 추가
-    if (userInfoStore.userInfoList.length === 0 && showLoginAlert) {
-      Swal.fire({
-        icon: 'warning',
-        title: '로그인이 필요합니다!',
-        text: '로그인을 하고 일기를 작성해 주세요! 😜',
-        confirmButtonText: '확인',
-        allowOutsideClick: false,
-      }).then(() => {
-        navigate('/login');
-      });
-    }
-  }, [userInfoStore.userInfoList.length, navigate, showLoginAlert]);
-  useEffect(() => {
-    setShowLoginAlert(true);
-  }, []);
 
   // userInfoList 배열에서 첫 번째 사용자 정보를 가져옴
   const user =
@@ -64,28 +32,8 @@ const NavigateBar = () => {
     };
   }, []);
 
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (
-        notifyMenuRef.current &&
-        !notifyMenuRef.current.contains(event.target)
-      ) {
-        setIsNotifyMenuOpen(false);
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, []);
-
   const handleProfArrowClick = () => {
     setIsProfMenuOpen(!isProfMenuOpen);
-  };
-
-  const handleNotifyArrowClick = () => {
-    setIsNotifyMenuOpen(!isNotifyMenuOpen);
   };
 
   return (
